explicitly define all imports
R > 3.3.0 requires all function calls to be correctly imported, even
from packages which are installed by default. This means the previous
density() and sd() calls now raise notes when the package is checked.
The stats and graphics packages are now imported, and the function
calls are now explicitly called e.g. stats::density()